daemon: cut over reconcile runtime actor lane
