+++ crates/atm-daemon/src/reconcile_runtime.rs
@@ -1,3 +1,4 @@
+use arc_swap::ArcSwap;
 use atm_core::boundary::{
     InboxIngress, NotificationSink, ReconcileRequest, ReconcileResult, WatchEventSource,
     WatchSubscriptionRequest,
@@ -6,13 +7,16 @@ use atm_core::error::AtmError;
 use atm_core::protocol::{NotificationEvent, NotificationKind, ProtocolErrorEnvelope};
 use std::collections::{HashMap, HashSet, VecDeque};
 use std::path::PathBuf;
-use std::sync::{Arc, Condvar, Mutex, mpsc};
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::mpsc::{self, Receiver, SyncSender, TrySendError};
+use std::sync::{Arc, Mutex};
 use std::thread::{self, JoinHandle};
-use std::time::Duration;
+use std::time::{Duration, Instant};
 
 use crate::worker_support::JoinHandleOwner;
 use crate::{DaemonSubsystem, SubsystemObservability};
 
+const DEFAULT_RECONCILE_QUEUE_CAPACITY: usize = 64;
 const DEFAULT_RECONCILE_DEBOUNCE: Duration = Duration::from_millis(25);
 const DEFAULT_RECONCILE_COMPLETION_TIMEOUT: Duration = Duration::from_secs(5);
 const DEFAULT_RECONCILE_IDLE_INTERVAL: Duration = Duration::from_millis(50);
@@ -25,73 +29,44 @@ const RECONCILE_SHUTDOWN_DEADLINE: Duration = Duration::from_secs(2);
 #[cfg(test)]
 const RECONCILE_SHUTDOWN_DEADLINE: Duration = Duration::from_millis(100);
 
+#[derive(Debug, Clone, Default)]
+pub(crate) struct ReconcileRuntimeStatus {
+    started: bool,
+    shutdown_requested: bool,
+    shutdown_started_at: Option<Instant>,
+    degraded_message: Option<Arc<str>>,
+}
+
 #[derive(Clone)]
 pub(crate) struct ReconcileRuntime {
     inner: Arc<ReconcileRuntimeInner>,
 }
 
-type ReconcileExecutor =
-    Arc<dyn Fn(&ReconcileRequest) -> Result<ReconcileResult, AtmError> + Send + Sync>;
-
 struct ReconcileRuntimeInner {
-    // Worker thread, reconcile callers, and shutdown path all access state
-    // concurrently; Mutex+Condvar still guards lifecycle and pending
-    // coalescing state in Y.21 while the actor command/reply contract is
-    // frozen ahead of the Y.22 production cutover.
-    state: Mutex<ReconcileState>,
-    wake: Condvar,
-    #[cfg(test)]
-    pending_changed: Condvar,
+    command_tx: SyncSender<ReconcileCommand>,
+    command_rx: Mutex<Option<Receiver<ReconcileCommand>>>,
+    status: Arc<ArcSwap<ReconcileRuntimeStatus>>,
+    worker: Arc<JoinHandleOwner>,
     debounce: Duration,
     executor: ReconcileExecutor,
+    notification_sink: Arc<dyn NotificationSink + Send + Sync>,
+    shutdown_deadline: Duration,
     observability: SubsystemObservability,
-    worker: Arc<JoinHandleOwner>,
-    #[cfg_attr(not(test), allow(dead_code))]
-    // Lock ordering invariant: if code ever needs both runtime state and this
-    // fingerprint registry, it must drop `state` before locking the registry
-    // mutex. The registry remains an ancillary side registry in Y.21 and must
-    // never nest inside the primary runtime lifecycle lock. Y.22 owns the
-    // final move into worker-owned reconcile state.
-    notification_fingerprints: Arc<Mutex<NotificationFingerprintRegistry>>,
+    start_claimed: AtomicBool,
 }
 
-#[derive(Default)]
-struct ReconcileState {
-    started: bool,
-    shutdown: bool,
-    worker_state: ReconcileWorkerState,
-}
+type ReconcileExecutor =
+    Arc<dyn Fn(&ReconcileRequest) -> Result<ReconcileExecution, AtmError> + Send + Sync>;
 
-impl ReconcileState {
-    #[cfg(test)]
-    fn counts(&self) -> (usize, usize, usize) {
-        (
-            self.worker_state.pending.len(),
-            self.worker_state.pending_order.len(),
-            0,
-        )
-    }
-
-    fn waiter_count(&self) -> usize {
-        self.worker_state
-            .pending
-            .values()
-            .map(|pending| pending.replies.len())
-            .sum()
-    }
-}
-
-#[derive(Default)]
-struct ReconcileWorkerState {
-    pending_epoch: u64,
-    pending: HashMap<ReconcileKey, PendingReconcile>,
-    pending_order: VecDeque<ReconcileKey>,
+pub(crate) struct ReconcileExecution {
+    result: ReconcileResult,
+    current_fingerprints: Option<HashSet<String>>,
 }
 
 pub(crate) enum ReconcileCommand {
     Reconcile {
         request: ReconcileRequest,
-        reply_tx: mpsc::SyncSender<Result<ReconcileResult, AtmError>>,
+        reply_tx: SyncSender<Result<ReconcileResult, AtmError>>,
     },
 }
 
@@ -104,7 +79,30 @@ struct ReconcileKey {
 
 struct PendingReconcile {
     request: ReconcileRequest,
-    replies: Vec<mpsc::SyncSender<Result<ReconcileResult, AtmError>>>,
+    replies: Vec<SyncSender<Result<ReconcileResult, AtmError>>>,
+}
+
+#[derive(Default)]
+struct ReconcileWorkerState {
+    pending_epoch: u64,
+    pending: HashMap<ReconcileKey, PendingReconcile>,
+    pending_order: VecDeque<ReconcileKey>,
+    notification_fingerprints: NotificationFingerprintRegistry,
+}
+
+impl ReconcileWorkerState {
+    fn waiter_count(&self) -> usize {
+        self.pending
+            .values()
+            .map(|pending| pending.replies.len())
+            .sum()
+    }
+}
+
+#[derive(Default)]
+struct NotificationFingerprintRegistry {
+    entries: HashMap<ReconcileKey, HashSet<String>>,
+    order: VecDeque<ReconcileKey>,
 }
 
 #[derive(Clone)]
@@ -133,15 +131,6 @@ impl ReconcileReplyErrorSnapshot {
         }
         .into_atm_error()
     }
-}
-
-#[derive(Default)]
-struct NotificationFingerprintRegistry {
-    // Each key retains the distinct fingerprint strings already emitted for one
-    // reconcile target so duplicate inbox projections do not re-notify until a
-    // genuinely new fingerprint appears.
-    entries: HashMap<ReconcileKey, HashSet<String>>,
-    order: VecDeque<ReconcileKey>,
 }
 
 impl ReconcileKey {
@@ -175,13 +164,7 @@ impl ReconcileRuntime {
         notification_sink: Arc<dyn NotificationSink + Send + Sync>,
         observability: SubsystemObservability,
     ) -> Self {
-        // Fingerprints must survive across executor invocations so duplicate
-        // reconcile cycles can compare the newest inbox projection with the
-        // previous one before deciding whether to emit a notification.
-        let notification_fingerprints =
-            Arc::new(Mutex::new(NotificationFingerprintRegistry::default()));
-        let notification_fingerprints_for_executor = Arc::clone(&notification_fingerprints);
-        Self::new_with_executor(
+        Self::new_with_executor_and_sink(
             Arc::new(move |request| {
                 let batch = watch_source.poll(WatchSubscriptionRequest {
                     home_dir: request.home_dir.clone(),
@@ -195,73 +178,71 @@ impl ReconcileRuntime {
                         agent: request.agent.clone(),
                     },
                 )?;
-                if should_emit_reconcile_notification(
-                    request,
-                    &import,
-                    inbox_ingress.as_ref(),
-                    notification_fingerprints_for_executor.as_ref(),
-                )? {
-                    notification_sink.deliver(NotificationEvent {
-                        kind: NotificationKind::ReconcileComplete,
-                        detail: format!(
-                            "observed_paths={} imported_sources={}",
-                            batch.paths.len(),
-                            import.source_files.len()
-                        ),
-                        team: Some(request.team.clone()),
-                        agent: Some(request.agent.clone()),
-                    })?;
-                }
-                Ok(ReconcileResult {
-                    observed_paths: batch.paths.len(),
-                    imported_sources: import.source_files.len(),
+                Ok(ReconcileExecution {
+                    result: ReconcileResult {
+                        observed_paths: batch.paths.len(),
+                        imported_sources: import.source_files.len(),
+                    },
+                    current_fingerprints: compute_reconcile_notification_fingerprints(
+                        &import,
+                        inbox_ingress.as_ref(),
+                    )?,
                 })
             }),
+            notification_sink,
+            DEFAULT_RECONCILE_QUEUE_CAPACITY,
             DEFAULT_RECONCILE_DEBOUNCE,
-            notification_fingerprints,
+            RECONCILE_SHUTDOWN_DEADLINE,
             observability,
         )
     }
 
-    fn new_with_executor(
+    fn new_with_executor_and_sink(
         executor: ReconcileExecutor,
+        notification_sink: Arc<dyn NotificationSink + Send + Sync>,
+        queue_capacity: usize,
         debounce: Duration,
-        notification_fingerprints: Arc<Mutex<NotificationFingerprintRegistry>>,
+        shutdown_deadline: Duration,
         observability: SubsystemObservability,
     ) -> Self {
+        let (command_tx, command_rx) = mpsc::sync_channel(queue_capacity);
         Self {
             inner: Arc::new(ReconcileRuntimeInner {
-                state: Mutex::new(ReconcileState::default()),
-                wake: Condvar::new(),
-                #[cfg(test)]
-                pending_changed: Condvar::new(),
+                command_tx,
+                command_rx: Mutex::new(Some(command_rx)),
+                status: Arc::new(ArcSwap::from_pointee(ReconcileRuntimeStatus::default())),
+                worker: Arc::new(JoinHandleOwner::default()),
                 debounce,
                 executor,
+                notification_sink,
+                shutdown_deadline,
                 observability,
-                worker: Arc::new(JoinHandleOwner::default()),
-                notification_fingerprints,
+                start_claimed: AtomicBool::new(false),
             }),
         }
     }
 
     pub(crate) fn start(&self) -> Result<(), AtmError> {
-        let mut state = self.inner.state.lock().map_err(|_| {
-            AtmError::daemon_unavailable("reconcile runtime state lock poisoned").with_recovery(
-                "Restart the daemon; reconcile lifecycle state can no longer be trusted.",
-            )
-        })?;
-        if state.started {
+        if self.inner.start_claimed.swap(true, Ordering::AcqRel) {
             return Ok(());
         }
-        state.started = true;
-        state.shutdown = false;
-        drop(state);
+
+        let Some(command_rx) = self.inner.take_command_rx()? else {
+            self.inner.start_claimed.store(false, Ordering::Release);
+            return Err(AtmError::daemon_unavailable(
+                "reconcile runtime command receiver was unavailable during startup",
+            )
+            .with_recovery(
+                "Restart atm-daemon; the reconcile worker lane could not claim its command receiver.",
+            ));
+        };
 
         let inner = Arc::clone(&self.inner);
         let handle = thread::Builder::new()
             .name("atm-daemon-reconcile".to_string())
-            .spawn(move || reconcile_worker_loop(inner))
+            .spawn(move || reconcile_worker_loop(inner, command_rx))
             .map_err(|source| {
+                self.inner.start_claimed.store(false, Ordering::Release);
                 self.inner.observability.emit_or_warn(
                     "start",
                     "failed",
@@ -270,21 +251,11 @@ impl ReconcileRuntime {
                 AtmError::daemon_unavailable("failed to spawn reconcile runtime worker")
                     .with_source(source)
             })?;
-        let mut state = match self.inner.state.lock() {
-            Ok(state) => state,
-            Err(_) => {
-                let _ = handle.join();
-                return Err(
-                    AtmError::daemon_unavailable("reconcile runtime state lock poisoned")
-                        .with_recovery(
-                            "Restart the daemon; reconcile lifecycle state can no longer be trusted.",
-                        ),
-                );
-            }
-        };
+
         self.inner.worker.install(handle).inspect_err(|_| {
-            state.shutdown = true;
+            self.inner.start_claimed.store(false, Ordering::Release);
         })?;
+        self.inner.mark_started();
         self.inner
             .observability
             .emit_or_warn("start", "ok", "reconcile runtime worker started");
@@ -292,191 +263,96 @@ impl ReconcileRuntime {
     }
 
     pub(crate) fn shutdown(&self) -> Result<(), AtmError> {
-        let handle = {
-            let mut state = self.inner.state.lock().map_err(|_| {
-                AtmError::daemon_unavailable("reconcile runtime state lock poisoned").with_recovery(
-                    "Restart the daemon; reconcile lifecycle state can no longer be trusted.",
-                )
-            })?;
-            state.shutdown = true;
-            state.worker_state = ReconcileWorkerState::default();
-            self.inner.wake.notify_all();
-            self.inner.worker.take()?
+        let Some(handle) = self.take_worker_for_shutdown()? else {
+            self.inner.mark_shutdown_requested();
+            self.inner.mark_worker_stopped();
+            return Ok(());
         };
-        if let Some(handle) = handle {
-            let worker_thread_id = handle.thread().id();
-            let (result_rx, join_helper) = spawn_reconcile_join_helper(handle)?;
-            self.complete_reconcile_shutdown(result_rx, join_helper, worker_thread_id)?;
-        }
-        Ok(())
-    }
-
-    fn complete_reconcile_shutdown(
-        &self,
-        result_rx: mpsc::Receiver<thread::Result<()>>,
-        join_helper: JoinHandle<()>,
-        worker_thread_id: thread::ThreadId,
-    ) -> Result<(), AtmError> {
-        match wait_for_reconcile_join(result_rx) {
-            Ok(Ok(())) => {
-                let _ = join_helper.join();
-                self.inner.observability.emit_or_warn(
-                    "shutdown",
-                    "ok",
-                    "reconcile runtime worker shut down cleanly",
-                );
-            }
-            Ok(Err(_)) => {
-                let _ = join_helper.join();
-                self.inner.observability.emit_or_warn(
-                    "shutdown",
-                    "failed",
-                    "reconcile runtime worker panicked during shutdown",
-                );
-                return Err(AtmError::daemon_unavailable(
-                    "reconcile runtime worker panicked during shutdown",
-                )
-                .with_recovery(
-                    "Restart atm-daemon; the reconcile background lane crashed while shutting down.",
-                ));
-            }
-            Err(ReconcileJoinStatus::Timeout) => {
-                drop(join_helper);
-                tracing::warn!(
-                    subsystem = "reconcile",
-                    action = "shutdown_detach",
-                    outcome = "deadline_exceeded",
-                    thread_id = ?worker_thread_id,
-                    timeout_ms = RECONCILE_SHUTDOWN_DEADLINE.as_millis(),
-                    "reconcile runtime worker exceeded shutdown deadline; detaching join helper"
-                );
-                self.inner.observability.emit_or_warn(
-                    "shutdown",
-                    "degraded",
-                    "reconcile runtime worker exceeded its shutdown deadline",
-                );
-                return Err(AtmError::daemon_unavailable(
-                    "reconcile runtime worker exceeded the bounded shutdown deadline",
-                )
-                .with_recovery(
-                    "Restart atm-daemon after the reconcile background lane becomes responsive again.",
-                ));
-            }
-            Err(ReconcileJoinStatus::Disconnected) => {
-                let _ = join_helper.join();
-                self.inner.observability.emit_or_warn(
-                    "shutdown",
-                    "failed",
-                    "reconcile runtime join helper disconnected during shutdown",
-                );
-                tracing::warn!(
-                    subsystem = "reconcile",
-                    action = "shutdown_join_helper",
-                    outcome = "disconnected",
-                    thread_id = ?worker_thread_id,
-                    "reconcile runtime worker join helper exited before reporting shutdown status"
-                );
-                return Err(AtmError::daemon_unavailable(
-                    "reconcile runtime join helper disconnected during shutdown",
-                )
-                .with_recovery(
-                    "Restart atm-daemon; the reconcile background lane did not report bounded shutdown status cleanly.",
-                ));
-            }
-        }
-        Ok(())
+        self.await_worker_shutdown(handle)
     }
 
     pub(crate) fn reconcile(&self, request: ReconcileRequest) -> Result<ReconcileResult, AtmError> {
         let request_team = request.team.clone();
         let request_agent = request.agent.clone();
-        let (reply_tx, reply_rx) = mpsc::sync_channel(1);
-        self.enqueue_reconcile_command(
-            ReconcileCommand::Reconcile { request, reply_tx },
-            &request_team,
-            &request_agent,
-        )?;
-        self.wait_for_reconcile_result(reply_rx, &request_team, &request_agent)
-    }
-
-    fn enqueue_reconcile_command(
-        &self,
-        command: ReconcileCommand,
-        request_team: &atm_core::types::TeamName,
-        request_agent: &atm_core::types::AgentName,
-    ) -> Result<(), AtmError> {
-        let mut state = self.inner.state.lock().map_err(|_| {
-            AtmError::daemon_unavailable("reconcile runtime state lock poisoned").with_recovery(
-                "Restart the daemon; reconcile lifecycle state can no longer be trusted.",
-            )
-        })?;
-        self.validate_reconcile_runtime_state(&state, request_team, request_agent)?;
-        match command {
-            ReconcileCommand::Reconcile { request, reply_tx } => {
-                state.worker_state.pending_epoch =
-                    state.worker_state.pending_epoch.saturating_add(1);
-                let key = ReconcileKey::from_request(&request);
-                if let Some(pending) = state.worker_state.pending.get_mut(&key) {
-                    pending.request = request;
-                    pending.replies.push(reply_tx);
-                } else {
-                    state.worker_state.pending_order.push_back(key.clone());
-                    state.worker_state.pending.insert(
-                        key,
-                        PendingReconcile {
-                            request,
-                            replies: vec![reply_tx],
-                        },
-                    );
-                }
-            }
-        }
-        self.inner.wake.notify_one();
-        #[cfg(test)]
-        self.inner.pending_changed.notify_all();
-        Ok(())
-    }
-
-    fn validate_reconcile_runtime_state(
-        &self,
-        state: &ReconcileState,
-        request_team: &atm_core::types::TeamName,
-        request_agent: &atm_core::types::AgentName,
-    ) -> Result<(), AtmError> {
-        if !state.started {
+        let status = self.inner.status_snapshot();
+        if !status.started {
             return Err(self.reconcile_unavailable_error(
                 "rejected",
                 "reconcile runtime is unavailable before daemon startup",
-                request_team,
-                request_agent,
+                &request_team,
+                &request_agent,
             ));
         }
-        if state.shutdown {
+        if status.shutdown_requested {
             return Err(self.reconcile_unavailable_error(
                 "rejected",
                 "reconcile runtime is unavailable during daemon shutdown",
-                request_team,
-                request_agent,
+                &request_team,
+                &request_agent,
             ));
         }
-        if state.waiter_count() >= MAX_RECONCILE_WAITERS {
-            return Err(self
-                .reconcile_unavailable_error(
+        if let Some(message) = &status.degraded_message {
+            return Err(
+                AtmError::daemon_unavailable(message.as_ref()).with_recovery(
+                    "Restart atm-daemon; the reconcile runtime worker lane is degraded.",
+                ),
+            );
+        }
+
+        let (reply_tx, reply_rx) = mpsc::sync_channel(1);
+        match self
+            .inner
+            .command_tx
+            .try_send(ReconcileCommand::Reconcile { request, reply_tx })
+        {
+            Ok(()) => {}
+            Err(TrySendError::Full(_)) => {
+                self.inner.observability.emit_or_warn(
+                    "reconcile",
                     "rejected",
-                    "reconcile runtime hit its concurrent waiter capacity",
-                    request_team,
-                    request_agent,
+                    "reconcile runtime command queue is full",
+                );
+                return Err(
+                    AtmError::daemon_unavailable(
+                        "reconcile runtime command queue is full; requests are backpressured",
+                    )
+                    .with_recovery(
+                        "Retry after earlier reconcile requests complete or reduce concurrent reconcile load.",
+                    ),
+                );
+            }
+            Err(TrySendError::Disconnected(_)) => {
+                let latest_status = self.inner.status_snapshot();
+                if let Some(message) = &latest_status.degraded_message {
+                    return Err(
+                        AtmError::daemon_unavailable(message.as_ref()).with_recovery(
+                            "Restart atm-daemon; the reconcile runtime worker lane is degraded.",
+                        ),
+                    );
+                }
+                if latest_status.shutdown_requested {
+                    return Err(self.reconcile_unavailable_error(
+                        "rejected",
+                        "reconcile runtime is unavailable during daemon shutdown",
+                        &request_team,
+                        &request_agent,
+                    ));
+                }
+                return Err(AtmError::daemon_unavailable(
+                    "reconcile runtime command channel is unavailable",
                 )
                 .with_recovery(
-                    "Reduce concurrent reconcile waiters or retry after earlier reconcile requests complete.",
+                    "Restart atm-daemon; the reconcile worker lane is no longer receiving commands.",
                 ));
+            }
         }
-        Ok(())
+
+        self.wait_for_reconcile_result(reply_rx, &request_team, &request_agent)
     }
 
     fn wait_for_reconcile_result(
         &self,
-        reply_rx: mpsc::Receiver<Result<ReconcileResult, AtmError>>,
+        reply_rx: Receiver<Result<ReconcileResult, AtmError>>,
         request_team: &atm_core::types::TeamName,
         request_agent: &atm_core::types::AgentName,
     ) -> Result<ReconcileResult, AtmError> {
@@ -516,111 +392,180 @@ impl ReconcileRuntime {
         )
     }
 
+    fn take_worker_for_shutdown(&self) -> Result<Option<JoinHandle<()>>, AtmError> {
+        self.inner.mark_shutdown_requested();
+        self.inner.worker.take()
+    }
+
+    fn await_worker_shutdown(&self, handle: JoinHandle<()>) -> Result<(), AtmError> {
+        let worker_thread_id = handle.thread().id();
+        let (join_helper, result_rx) = spawn_reconcile_join_helper(handle)?;
+        match result_rx.recv_timeout(self.inner.shutdown_deadline) {
+            Ok(Ok(())) => self.handle_shutdown_joined(join_helper),
+            Ok(Err(_)) => self.handle_shutdown_panic(join_helper),
+            Err(mpsc::RecvTimeoutError::Timeout) => {
+                self.handle_shutdown_timeout(join_helper, worker_thread_id)
+            }
+            Err(mpsc::RecvTimeoutError::Disconnected) => {
+                self.handle_shutdown_disconnect(join_helper, worker_thread_id)
+            }
+        }
+    }
+
+    fn handle_shutdown_joined(&self, join_helper: JoinHandle<()>) -> Result<(), AtmError> {
+        let _ = join_helper.join();
+        self.inner.mark_worker_stopped();
+        self.inner.observability.emit_or_warn(
+            "shutdown",
+            "ok",
+            "reconcile runtime worker shut down cleanly",
+        );
+        Ok(())
+    }
+
+    fn handle_shutdown_panic(&self, join_helper: JoinHandle<()>) -> Result<(), AtmError> {
+        let _ = join_helper.join();
+        self.inner
+            .mark_worker_degraded("reconcile runtime worker panicked during shutdown");
+        self.inner.observability.emit_or_warn(
+            "shutdown",
+            "failed",
+            "reconcile runtime worker panicked during shutdown",
+        );
+        Err(AtmError::daemon_unavailable(
+            "reconcile runtime worker panicked during shutdown",
+        )
+        .with_recovery(
+            "Restart atm-daemon; the reconcile background lane crashed while shutting down.",
+        ))
+    }
+
+    fn handle_shutdown_timeout(
+        &self,
+        join_helper: JoinHandle<()>,
+        worker_thread_id: thread::ThreadId,
+    ) -> Result<(), AtmError> {
+        self.inner.observability.emit_or_warn(
+            "shutdown",
+            "degraded",
+            "reconcile runtime worker exceeded its shutdown deadline",
+        );
+        tracing::warn!(
+            subsystem = "reconcile",
+            action = "shutdown_detach",
+            outcome = "deadline_exceeded",
+            thread_id = ?worker_thread_id,
+            timeout_ms = self.inner.shutdown_deadline.as_millis(),
+            "reconcile runtime worker exceeded shutdown deadline; detaching join helper"
+        );
+        drop(join_helper);
+        Err(AtmError::daemon_unavailable(
+            "reconcile runtime worker exceeded the bounded shutdown deadline",
+        )
+        .with_recovery(
+            "Restart atm-daemon after the reconcile background lane becomes responsive again.",
+        ))
+    }
+
+    fn handle_shutdown_disconnect(
+        &self,
+        join_helper: JoinHandle<()>,
+        worker_thread_id: thread::ThreadId,
+    ) -> Result<(), AtmError> {
+        let _ = join_helper.join();
+        self.inner
+            .mark_worker_degraded("reconcile runtime join helper disconnected during shutdown");
+        self.inner.observability.emit_or_warn(
+            "shutdown",
+            "failed",
+            "reconcile runtime join helper disconnected during shutdown",
+        );
+        tracing::warn!(
+            subsystem = "reconcile",
+            action = "shutdown_join_helper",
+            outcome = "disconnected",
+            thread_id = ?worker_thread_id,
+            "reconcile runtime join helper exited before reporting shutdown status"
+        );
+        Err(AtmError::daemon_unavailable(
+            "reconcile runtime join helper disconnected during shutdown",
+        )
+        .with_recovery(
+            "Restart atm-daemon; the reconcile background lane did not report bounded shutdown status cleanly.",
+        ))
+    }
+
     #[cfg(test)]
     pub(crate) fn new_for_test(executor: ReconcileExecutor, debounce: Duration) -> Self {
-        Self::new_with_executor(
+        Self::new_with_executor_and_sink(
             executor,
+            Arc::new(TestNotificationSink),
+            DEFAULT_RECONCILE_QUEUE_CAPACITY,
             debounce,
-            Arc::new(Mutex::new(NotificationFingerprintRegistry::default())),
+            RECONCILE_SHUTDOWN_DEADLINE,
             SubsystemObservability::disabled(DaemonSubsystem::ReconcileRuntime),
         )
     }
+}
 
-    #[cfg(test)]
-    pub(crate) fn state_counts_for_test(&self) -> (usize, usize, usize) {
-        self.inner.state.lock().expect("state lock").counts()
+impl ReconcileRuntimeInner {
+    fn take_command_rx(&self) -> Result<Option<Receiver<ReconcileCommand>>, AtmError> {
+        let mut receiver = self.command_rx.lock().map_err(|_| {
+            AtmError::daemon_unavailable("reconcile runtime receiver lock poisoned").with_recovery(
+                "Restart atm-daemon; reconcile command receiver ownership can no longer be trusted.",
+            )
+        })?;
+        Ok(receiver.take())
     }
 
-    #[cfg(test)]
-    pub(crate) fn wait_for_pending_count_for_test(
+    fn status_snapshot(&self) -> Arc<ReconcileRuntimeStatus> {
+        self.status.load_full()
+    }
+
+    fn publish_status(
         &self,
-        minimum_pending: usize,
-        timeout: Duration,
-    ) -> bool {
-        let deadline = std::time::Instant::now() + timeout;
-        let mut state = self.inner.state.lock().expect("state lock");
-        while state.worker_state.pending.len() < minimum_pending {
-            let now = std::time::Instant::now();
-            if now >= deadline {
-                return false;
-            }
-            let wait = self
-                .inner
-                .pending_changed
-                .wait_timeout(state, deadline.saturating_duration_since(now))
-                .expect("pending change wait");
-            state = wait.0;
-            if wait.1.timed_out() {
-                return state.worker_state.pending.len() >= minimum_pending;
-            }
-        }
-        true
+        mutate: impl FnOnce(&ReconcileRuntimeStatus) -> ReconcileRuntimeStatus,
+    ) {
+        let current = self.status.load_full();
+        self.status.store(Arc::new(mutate(current.as_ref())));
     }
 
-    #[cfg(test)]
-    pub(crate) fn wait_for_pending_waiter_count_for_test(
-        &self,
-        minimum_waiters: usize,
-        timeout: Duration,
-    ) -> bool {
-        let deadline = std::time::Instant::now() + timeout;
-        let mut state = self.inner.state.lock().expect("state lock");
-        while state.waiter_count() < minimum_waiters {
-            let now = std::time::Instant::now();
-            if now >= deadline {
-                return false;
-            }
-            let wait = self
-                .inner
-                .pending_changed
-                .wait_timeout(state, deadline.saturating_duration_since(now))
-                .expect("pending change wait");
-            state = wait.0;
-            if wait.1.timed_out() {
-                return state.waiter_count() >= minimum_waiters;
-            }
-        }
-        true
+    fn mark_started(&self) {
+        self.status.store(Arc::new(ReconcileRuntimeStatus {
+            started: true,
+            shutdown_requested: false,
+            shutdown_started_at: None,
+            degraded_message: None,
+        }));
     }
 
-    #[cfg(test)]
-    pub(crate) fn wait_for_pending_agent_for_test(&self, agent: &str, timeout: Duration) -> bool {
-        let deadline = std::time::Instant::now() + timeout;
-        let mut state = self.inner.state.lock().expect("state lock");
-        while !state
-            .worker_state
-            .pending
-            .values()
-            .any(|pending| pending.request.agent.as_str() == agent)
-        {
-            let now = std::time::Instant::now();
-            if now >= deadline {
-                return false;
-            }
-            let wait = self
-                .inner
-                .pending_changed
-                .wait_timeout(state, deadline.saturating_duration_since(now))
-                .expect("pending change wait");
-            state = wait.0;
-            if wait.1.timed_out() {
-                return state
-                    .worker_state
-                    .pending
-                    .values()
-                    .any(|pending| pending.request.agent.as_str() == agent);
-            }
-        }
-        true
+    fn mark_shutdown_requested(&self) {
+        self.publish_status(|status| ReconcileRuntimeStatus {
+            shutdown_requested: true,
+            shutdown_started_at: status.shutdown_started_at.or(Some(Instant::now())),
+            ..status.clone()
+        });
+    }
+
+    fn mark_worker_stopped(&self) {
+        self.publish_status(|status| ReconcileRuntimeStatus {
+            degraded_message: None,
+            ..status.clone()
+        });
+    }
+
+    fn mark_worker_degraded(&self, message: &'static str) {
+        self.publish_status(|status| ReconcileRuntimeStatus {
+            degraded_message: Some(Arc::<str>::from(message)),
+            ..status.clone()
+        });
     }
 }
 
-fn should_emit_reconcile_notification(
-    request: &ReconcileRequest,
+fn compute_reconcile_notification_fingerprints(
     import: &atm_core::boundary::InboxIngressImportResponse,
     inbox_ingress: &dyn InboxIngress,
-    notification_fingerprints: &Mutex<NotificationFingerprintRegistry>,
-) -> Result<bool, AtmError> {
+) -> Result<Option<HashSet<String>>, AtmError> {
     let mut current_fingerprints = HashSet::new();
     for source in &import.source_files {
         for message in &source.messages {
@@ -632,11 +577,169 @@ fn should_emit_reconcile_notification(
                 )
                 .fingerprint;
             let Some(fingerprint) = fingerprint else {
-                return Ok(true);
+                return Ok(None);
             };
             current_fingerprints.insert(fingerprint);
         }
     }
+    Ok(Some(current_fingerprints))
+}
+
+fn reconcile_worker_loop(
+    inner: Arc<ReconcileRuntimeInner>,
+    command_rx: Receiver<ReconcileCommand>,
+) {
+    let mut worker_state = ReconcileWorkerState::default();
+    loop {
+        if inner.status_snapshot().shutdown_requested {
+            inner.mark_worker_stopped();
+            return;
+        }
+
+        let first_command = match command_rx.recv_timeout(DEFAULT_RECONCILE_IDLE_INTERVAL) {
+            Ok(command) => command,
+            Err(mpsc::RecvTimeoutError::Timeout) => continue,
+            Err(mpsc::RecvTimeoutError::Disconnected) => {
+                inner.mark_worker_stopped();
+                return;
+            }
+        };
+        enqueue_pending_reconcile_command(inner.as_ref(), &mut worker_state, first_command);
+
+        let pending = match debounce_pending_reconcile_batch(
+            inner.as_ref(),
+            &command_rx,
+            &mut worker_state,
+        ) {
+            Some(pending) => pending,
+            None => {
+                inner.mark_worker_stopped();
+                return;
+            }
+        };
+
+        for pending_request in pending {
+            let outcome = execute_reconcile_request(
+                inner.as_ref(),
+                &mut worker_state,
+                &pending_request.request,
+            );
+            if record_reconcile_outcome(pending_request, outcome).is_none() {
+                inner.mark_worker_stopped();
+                return;
+            }
+        }
+    }
+}
+
+fn enqueue_pending_reconcile_command(
+    inner: &ReconcileRuntimeInner,
+    worker_state: &mut ReconcileWorkerState,
+    command: ReconcileCommand,
+) {
+    match command {
+        ReconcileCommand::Reconcile { request, reply_tx } => {
+            if worker_state.waiter_count() >= MAX_RECONCILE_WAITERS {
+                let _ = reply_tx.send(Err(AtmError::daemon_unavailable(
+                        "reconcile runtime hit its concurrent waiter capacity",
+                    )
+                    .with_recovery(
+                        "Reduce concurrent reconcile waiters or retry after earlier reconcile requests complete.",
+                    )));
+                inner.observability.emit_or_warn(
+                    "reconcile",
+                    "rejected",
+                    "reconcile runtime hit its concurrent waiter capacity",
+                );
+                return;
+            }
+            worker_state.pending_epoch = worker_state.pending_epoch.saturating_add(1);
+            let key = ReconcileKey::from_request(&request);
+            if let Some(pending) = worker_state.pending.get_mut(&key) {
+                pending.request = request;
+                pending.replies.push(reply_tx);
+            } else {
+                worker_state.pending_order.push_back(key.clone());
+                worker_state.pending.insert(
+                    key,
+                    PendingReconcile {
+                        request,
+                        replies: vec![reply_tx],
+                    },
+                );
+            }
+        }
+    }
+}
+
+fn debounce_pending_reconcile_batch(
+    inner: &ReconcileRuntimeInner,
+    command_rx: &Receiver<ReconcileCommand>,
+    worker_state: &mut ReconcileWorkerState,
+) -> Option<Vec<PendingReconcile>> {
+    let mut debounce_epoch = worker_state.pending_epoch;
+    let mut debounce_extensions = 0u32;
+    loop {
+        if inner.status_snapshot().shutdown_requested {
+            return None;
+        }
+        match command_rx.recv_timeout(inner.debounce) {
+            Ok(command) => {
+                enqueue_pending_reconcile_command(inner, worker_state, command);
+                if worker_state.pending_epoch != debounce_epoch {
+                    debounce_epoch = worker_state.pending_epoch;
+                    debounce_extensions = debounce_extensions.saturating_add(1);
+                    if debounce_extensions >= MAX_RECONCILE_DEBOUNCE_EXTENSIONS {
+                        break;
+                    }
+                }
+            }
+            Err(mpsc::RecvTimeoutError::Timeout) => break,
+            Err(mpsc::RecvTimeoutError::Disconnected) => break,
+        }
+    }
+    Some(drain_pending_reconcile_batch(worker_state))
+}
+
+fn drain_pending_reconcile_batch(worker_state: &mut ReconcileWorkerState) -> Vec<PendingReconcile> {
+    let pending_order = std::mem::take(&mut worker_state.pending_order);
+    let mut drained = Vec::with_capacity(pending_order.len());
+    for key in pending_order {
+        if let Some(pending) = worker_state.pending.remove(&key) {
+            drained.push(pending);
+        }
+    }
+    drained
+}
+
+fn execute_reconcile_request(
+    inner: &ReconcileRuntimeInner,
+    worker_state: &mut ReconcileWorkerState,
+    request: &ReconcileRequest,
+) -> Result<ReconcileResult, AtmError> {
+    let execution = (inner.executor)(request)?;
+    if should_emit_reconcile_notification(worker_state, request, execution.current_fingerprints)? {
+        inner.notification_sink.deliver(NotificationEvent {
+            kind: NotificationKind::ReconcileComplete,
+            detail: format!(
+                "observed_paths={} imported_sources={}",
+                execution.result.observed_paths, execution.result.imported_sources
+            ),
+            team: Some(request.team.clone()),
+            agent: Some(request.agent.clone()),
+        })?;
+    }
+    Ok(execution.result)
+}
+
+fn should_emit_reconcile_notification(
+    worker_state: &mut ReconcileWorkerState,
+    request: &ReconcileRequest,
+    current_fingerprints: Option<HashSet<String>>,
+) -> Result<bool, AtmError> {
+    let Some(mut current_fingerprints) = current_fingerprints else {
+        return Ok(true);
+    };
 
     let key = ReconcileKey::from_request(request);
     if current_fingerprints.len() > MAX_RECONCILE_FINGERPRINTS_PER_KEY {
@@ -658,9 +761,8 @@ fn should_emit_reconcile_notification(
             "reconcile notification fingerprint set exceeded the per-key bounded cap; truncating deterministically"
         );
     }
-    let mut fingerprints = notification_fingerprints.lock().map_err(|_| {
-        AtmError::daemon_unavailable("reconcile notification fingerprint state lock poisoned")
-    })?;
+
+    let fingerprints = &mut worker_state.notification_fingerprints;
     let changed = fingerprints
         .entries
         .get(&key)
@@ -690,131 +792,7 @@ fn should_emit_reconcile_notification(
     Ok(changed)
 }
 
-fn reconcile_worker_loop(inner: Arc<ReconcileRuntimeInner>) {
-    loop {
-        let pending = match take_pending_reconcile_batch(inner.as_ref()) {
-            Some(pending) => pending,
-            None => return,
-        };
-
-        for pending_request in pending {
-            let outcome = match (inner.executor)(&pending_request.request) {
-                Ok(result) => Ok(result),
-                Err(error) => {
-                    tracing::warn!(
-                        subsystem = "reconcile",
-                        action = "executor",
-                        outcome = "failed",
-                        team = %pending_request.request.team,
-                        agent = %pending_request.request.agent,
-                        %error,
-                        "reconcile runtime executor failed"
-                    );
-                    Err(error)
-                }
-            };
-            if record_reconcile_outcome(inner.as_ref(), pending_request, outcome).is_none() {
-                return;
-            }
-        }
-    }
-}
-
-enum ReconcileJoinStatus {
-    Timeout,
-    Disconnected,
-}
-
-fn spawn_reconcile_join_helper(
-    handle: JoinHandle<()>,
-) -> Result<(mpsc::Receiver<std::thread::Result<()>>, JoinHandle<()>), AtmError> {
-    let (result_tx, result_rx) = mpsc::sync_channel(1);
-    let join_helper = thread::Builder::new()
-        .name("atm-daemon-reconcile-join".to_string())
-        .spawn(move || {
-            let _ = result_tx.send(handle.join());
-        })
-        .map_err(|source| {
-            AtmError::daemon_unavailable(
-                "failed to spawn reconcile runtime join helper during shutdown",
-            )
-            .with_recovery(
-                "Restart atm-daemon; reconcile shutdown could not create its bounded join helper.",
-            )
-            .with_source(source)
-        })?;
-    Ok((result_rx, join_helper))
-}
-
-fn wait_for_reconcile_join(
-    result_rx: mpsc::Receiver<std::thread::Result<()>>,
-) -> Result<std::thread::Result<()>, ReconcileJoinStatus> {
-    result_rx
-        .recv_timeout(RECONCILE_SHUTDOWN_DEADLINE)
-        .map_err(|error| match error {
-            mpsc::RecvTimeoutError::Timeout => ReconcileJoinStatus::Timeout,
-            mpsc::RecvTimeoutError::Disconnected => ReconcileJoinStatus::Disconnected,
-        })
-}
-
-fn take_pending_reconcile_batch(inner: &ReconcileRuntimeInner) -> Option<Vec<PendingReconcile>> {
-    let mut state = inner.state.lock().ok()?;
-    while state.worker_state.pending.is_empty() && !state.shutdown {
-        let wait = inner
-            .wake
-            .wait_timeout(state, DEFAULT_RECONCILE_IDLE_INTERVAL)
-            .ok()?;
-        state = wait.0;
-    }
-    if state.shutdown {
-        return None;
-    }
-    debounce_pending_reconcile_batch(inner, state)
-}
-
-fn debounce_pending_reconcile_batch(
-    inner: &ReconcileRuntimeInner,
-    mut state: std::sync::MutexGuard<'_, ReconcileState>,
-) -> Option<Vec<PendingReconcile>> {
-    let mut debounce_epoch = state.worker_state.pending_epoch;
-    let mut debounce_extensions = 0u32;
-    loop {
-        let wait = inner.wake.wait_timeout(state, inner.debounce).ok()?;
-        state = wait.0;
-        if state.shutdown {
-            return None;
-        }
-        if state.worker_state.pending_epoch != debounce_epoch {
-            debounce_epoch = state.worker_state.pending_epoch;
-            debounce_extensions = debounce_extensions.saturating_add(1);
-            if debounce_extensions >= MAX_RECONCILE_DEBOUNCE_EXTENSIONS {
-                break;
-            }
-            continue;
-        }
-        if wait.1.timed_out() {
-            break;
-        }
-    }
-    let drained = drain_pending_reconcile_batch(&mut state);
-    #[cfg(test)]
-    inner.pending_changed.notify_all();
-    Some(drained)
-}
-
-fn drain_pending_reconcile_batch(state: &mut ReconcileState) -> Vec<PendingReconcile> {
-    let pending_order = std::mem::take(&mut state.worker_state.pending_order);
-    let mut drained = Vec::with_capacity(pending_order.len());
-    for key in pending_order {
-        if let Some(pending) = state.worker_state.pending.remove(&key) {
-            drained.push(pending);
-        }
-    }
-    drained
-}
-
 fn record_reconcile_outcome(
-    _inner: &ReconcileRuntimeInner,
     pending_request: PendingReconcile,
     outcome: Result<ReconcileResult, AtmError>,
 ) -> Option<()> {
@@ -833,629 +811,40 @@ fn clone_reconcile_outcome(
     }
 }
 
+fn spawn_reconcile_join_helper(
+    handle: JoinHandle<()>,
+) -> Result<(JoinHandle<()>, Receiver<std::thread::Result<()>>), AtmError> {
+    let (result_tx, result_rx) = mpsc::sync_channel(1);
+    let join_helper = thread::Builder::new()
+        .name("atm-daemon-reconcile-join".to_string())
+        .spawn(move || {
+            let _ = result_tx.send(handle.join());
+        })
+        .map_err(|source| {
+            AtmError::daemon_unavailable(
+                "failed to spawn reconcile runtime join helper during shutdown",
+            )
+            .with_recovery(
+                "Restart atm-daemon; reconcile shutdown could not create its bounded join helper.",
+            )
+            .with_source(source)
+        })?;
+    Ok((join_helper, result_rx))
+}
+
 #[cfg(test)]
-mod tests {
-    use super::{
-        MAX_RECONCILE_DEBOUNCE_EXTENSIONS, MAX_RECONCILE_FINGERPRINT_KEYS,
-        MAX_RECONCILE_FINGERPRINTS_PER_KEY, ReconcileKey, ReconcileRuntime,
-    };
-    use atm_core::boundary::{
-        self, InboxIngress, InboxIngressDiagnosticsRequest, InboxIngressDiagnosticsResponse,
-        InboxIngressIdentityFingerprintRequest, InboxIngressIdentityFingerprintResponse,
-        InboxIngressImportRequest, InboxIngressImportResponse, NotificationEvent, NotificationSink,
-        ReconcileRequest, WatchEventBatch, WatchEventSource, WatchSubscriptionRequest,
-    };
-    use atm_core::protocol::ReconcileResult;
-    use atm_core::roles::ROLE_TEAM_LEAD;
-    use atm_core::schema::{AtmMessageId, MessageEnvelope};
-    use atm_core::types::IsoTimestamp;
-    use chrono::Utc;
-    use serde_json::Map;
-    use std::sync::atomic::{AtomicU64, Ordering};
-    use std::sync::{Arc, Condvar, Mutex, mpsc};
-    use std::time::Duration;
+struct TestNotificationSink;
 
-    fn unique_home_dir() -> std::path::PathBuf {
-        static NEXT: AtomicU64 = AtomicU64::new(0);
-        std::env::temp_dir().join(format!(
-            "atm-reconcile-test-{}-{}",
-            std::process::id(),
-            NEXT.fetch_add(1, Ordering::Relaxed)
-        ))
-    }
+#[cfg(test)]
+impl atm_core::boundary::sealed::Sealed for TestNotificationSink {}
 
-    fn request() -> ReconcileRequest {
-        ReconcileRequest {
-            home_dir: unique_home_dir(),
-            team: "test-team".parse().expect("team"),
-            agent: "test-agent".parse().expect("agent"),
-        }
-    }
-
-    fn request_for(agent: &str) -> ReconcileRequest {
-        ReconcileRequest {
-            home_dir: unique_home_dir(),
-            team: "test-team".parse().expect("team"),
-            agent: agent.parse().expect("agent"),
-        }
-    }
-
-    #[test]
-    fn reconcile_runtime_actor_coalesces_identical_requests_into_one_worker_run() {
-        let calls = Arc::new(Mutex::new(0usize));
-        let runtime = ReconcileRuntime::new_for_test(
-            Arc::new({
-                let calls = Arc::clone(&calls);
-                move |_| {
-                    *calls.lock().expect("calls") += 1;
-                    Ok(ReconcileResult {
-                        observed_paths: 2,
-                        imported_sources: 1,
-                    })
-                }
-            }),
-            Duration::from_millis(200),
-        );
-        runtime.start().expect("start");
-
-        let runtime_a = runtime.clone();
-        let runtime_b = runtime.clone();
-        let request_a = request();
-        let request_b = request_a.clone();
-        let first = std::thread::spawn(move || runtime_a.reconcile(request_a).expect("first"));
-        assert!(
-            runtime.wait_for_pending_count_for_test(1, Duration::from_secs(1)),
-            "first reconcile request never entered the pending queue"
-        );
-        assert!(
-            runtime.wait_for_pending_waiter_count_for_test(1, Duration::from_secs(1)),
-            "first reconcile waiter never entered the pending queue"
-        );
-        let second = std::thread::spawn(move || runtime_b.reconcile(request_b).expect("second"));
-        assert!(
-            runtime.wait_for_pending_waiter_count_for_test(2, Duration::from_secs(1)),
-            "duplicate reconcile requests never shared the same pending work item"
-        );
-        assert_eq!(first.join().expect("join").observed_paths, 2);
-        assert_eq!(second.join().expect("join").imported_sources, 1);
-        assert_eq!(*calls.lock().expect("calls"), 1);
-        runtime.shutdown().expect("shutdown");
-    }
-
-    #[test]
-    fn reconcile_runtime_actor_fans_one_result_to_all_waiters_for_a_key() {
-        let calls = Arc::new(Mutex::new(0usize));
-        let runtime = ReconcileRuntime::new_for_test(
-            Arc::new({
-                let calls = Arc::clone(&calls);
-                move |_| {
-                    *calls.lock().expect("calls") += 1;
-                    Ok(ReconcileResult {
-                        observed_paths: 7,
-                        imported_sources: 3,
-                    })
-                }
-            }),
-            Duration::from_millis(200),
-        );
-        runtime.start().expect("start");
-
-        let request = request();
-        let runtime_a = runtime.clone();
-        let runtime_b = runtime.clone();
-        let request_a = request.clone();
-        let request_b = request;
-        let first = std::thread::spawn(move || runtime_a.reconcile(request_a).expect("first"));
-        assert!(
-            runtime.wait_for_pending_count_for_test(1, Duration::from_secs(1)),
-            "first reconcile request never entered the pending queue"
-        );
-        let second = std::thread::spawn(move || runtime_b.reconcile(request_b).expect("second"));
-        assert!(
-            runtime.wait_for_pending_waiter_count_for_test(2, Duration::from_secs(1)),
-            "duplicate reconcile requests never shared one worker-owned pending reply fanout"
-        );
-
-        let first_result = first.join().expect("join");
-        let second_result = second.join().expect("join");
-        assert_eq!(first_result, second_result);
-        assert_eq!(first_result.observed_paths, 7);
-        assert_eq!(first_result.imported_sources, 3);
-        assert_eq!(*calls.lock().expect("calls"), 1);
-        runtime.shutdown().expect("shutdown");
-    }
-
-    #[test]
-    fn reconcile_runtime_actor_preserves_bounded_debounce_extensions() {
-        let calls = Arc::new(Mutex::new(0usize));
-        let runtime = ReconcileRuntime::new_for_test(
-            Arc::new({
-                let calls = Arc::clone(&calls);
-                move |_| {
-                    *calls.lock().expect("calls") += 1;
-                    Ok(ReconcileResult {
-                        observed_paths: 1,
-                        imported_sources: 1,
-                    })
-                }
-            }),
-            Duration::from_millis(25),
-        );
-        runtime.start().expect("start");
-
-        let request = request();
-        let started = std::time::Instant::now();
-        let mut joins = Vec::new();
-        joins.push({
-            let runtime = runtime.clone();
-            let request = request.clone();
-            std::thread::spawn(move || runtime.reconcile(request).expect("first"))
-        });
-        assert!(
-            runtime.wait_for_pending_count_for_test(1, Duration::from_secs(1)),
-            "first reconcile request never entered the pending queue"
-        );
-        for _ in 0..=MAX_RECONCILE_DEBOUNCE_EXTENSIONS {
-            let runtime = runtime.clone();
-            let request = request.clone();
-            joins.push(std::thread::spawn(move || {
-                runtime.reconcile(request).expect("duplicate")
-            }));
-        }
-
-        for join in joins {
-            let result = join.join().expect("join");
-            assert_eq!(result.observed_paths, 1);
-            assert_eq!(result.imported_sources, 1);
-        }
-        assert_eq!(*calls.lock().expect("calls"), 1);
-        assert!(
-            started.elapsed() < Duration::from_secs(2),
-            "bounded debounce extensions never converged to one worker run"
-        );
-        runtime.shutdown().expect("shutdown");
-    }
-
-    #[test]
-    fn reconcile_runtime_cleans_up_pending_waiters_during_shutdown() {
-        let runtime = ReconcileRuntime::new_for_test(
-            Arc::new(|_| {
-                Ok(ReconcileResult {
-                    observed_paths: 1,
-                    imported_sources: 1,
-                })
-            }),
-            Duration::from_millis(250),
-        );
-        runtime.start().expect("start");
-
-        let runtime_for_thread = runtime.clone();
-        let join = std::thread::spawn(move || runtime_for_thread.reconcile(request()));
-        assert!(
-            runtime.wait_for_pending_count_for_test(1, Duration::from_secs(1)),
-            "reconcile request never entered the pending queue"
-        );
-        runtime.shutdown().expect("shutdown");
-
-        let error = join
-            .join()
-            .expect("join")
-            .expect_err("shutdown interruption");
-        assert!(error.message.contains("shut down before completion"));
-        assert_eq!(runtime.state_counts_for_test(), (0, 0, 0));
-    }
-
-    #[test]
-    fn reconcile_runtime_returns_executor_failures() {
-        let runtime = ReconcileRuntime::new_for_test(
-            Arc::new(|_| {
-                Err(atm_core::error::AtmError::daemon_unavailable(
-                    "reconcile failed",
-                ))
-            }),
-            Duration::from_millis(10),
-        );
-        runtime.start().expect("start");
-        let error = runtime.reconcile(request()).expect_err("failure");
-        assert!(error.message.contains("reconcile failed"));
-        runtime.shutdown().expect("shutdown");
-    }
-
-    #[test]
-    fn reconcile_runtime_preserves_trigger_order_and_signals_completion() {
-        let order = Arc::new(Mutex::new(Vec::new()));
-        let (started_tx, started_rx) = mpsc::channel();
-        let release = Arc::new((Mutex::new(false), Condvar::new()));
-        let runtime = ReconcileRuntime::new_for_test(
-            Arc::new({
-                let order = Arc::clone(&order);
-                let started_tx = started_tx.clone();
-                let release = Arc::clone(&release);
-                move |request| {
-                    order
-                        .lock()
-                        .expect("order")
-                        .push(request.agent.as_str().to_string());
-                    started_tx
-                        .send(request.agent.as_str().to_string())
-                        .expect("started");
-                    if request.agent.as_str() == "agent-a" {
-                        let (released, wake) = &*release;
-                        let mut released = released.lock().expect("released");
-                        while !*released {
-                            let wait = wake
-                                .wait_timeout(released, Duration::from_secs(1))
-                                .expect("wait release");
-                            released = wait.0;
-                            assert!(!wait.1.timed_out(), "agent-a release timed out");
-                        }
-                    }
-                    Ok(ReconcileResult {
-                        observed_paths: 1,
-                        imported_sources: 1,
-                    })
-                }
-            }),
-            Duration::from_millis(10),
-        );
-        runtime.start().expect("start");
-
-        let runtime_a = runtime.clone();
-        let runtime_b = runtime.clone();
-        let first = std::thread::spawn(move || runtime_a.reconcile(request_for("agent-a")));
-        assert_eq!(started_rx.recv().expect("first started"), "agent-a");
-        let second = std::thread::spawn(move || runtime_b.reconcile(request_for("agent-b")));
-        assert!(
-            runtime.wait_for_pending_agent_for_test("agent-b", Duration::from_secs(1)),
-            "agent-b never entered the pending queue before agent-a was released"
-        );
-        let (released, wake) = &*release;
-        *released.lock().expect("released") = true;
-        wake.notify_all();
-
-        let first_result = first.join().expect("first join").expect("first result");
-        let second_result = second.join().expect("second join").expect("second result");
-        assert_eq!(first_result.observed_paths, 1);
-        assert_eq!(second_result.imported_sources, 1);
-        assert_eq!(
-            order.lock().expect("order").as_slice(),
-            ["agent-a".to_string(), "agent-b".to_string()]
-        );
-        runtime.shutdown().expect("shutdown");
-    }
-
-    #[derive(Clone)]
-    struct FakeWatchSource;
-
-    impl boundary::sealed::Sealed for FakeWatchSource {}
-
-    impl WatchEventSource for FakeWatchSource {
-        fn poll(
-            &self,
-            _request: WatchSubscriptionRequest,
-        ) -> Result<WatchEventBatch, atm_core::error::AtmError> {
-            Ok(WatchEventBatch {
-                paths: vec![std::env::temp_dir().join("watch.json")],
-            })
-        }
-    }
-
-    #[derive(Clone)]
-    struct FakeInboxIngress {
-        imports: Arc<Mutex<Vec<InboxIngressImportResponse>>>,
-    }
-
-    impl FakeInboxIngress {
-        fn new(imports: Vec<InboxIngressImportResponse>) -> Self {
-            Self {
-                imports: Arc::new(Mutex::new(imports)),
-            }
-        }
-    }
-
-    impl boundary::sealed::Sealed for FakeInboxIngress {}
-
-    impl InboxIngress for FakeInboxIngress {
-        fn import_inbox_source(
-            &self,
-            _request: InboxIngressImportRequest,
-        ) -> Result<InboxIngressImportResponse, atm_core::error::AtmError> {
-            let mut imports = self.imports.lock().expect("imports");
-            if imports.is_empty() {
-                return Ok(InboxIngressImportResponse {
-                    source_files: Vec::new(),
-                });
-            }
-            Ok(imports.remove(0))
-        }
-
-        fn compute_identity_fingerprint(
-            &self,
-            request: InboxIngressIdentityFingerprintRequest,
-        ) -> InboxIngressIdentityFingerprintResponse {
-            InboxIngressIdentityFingerprintResponse {
-                fingerprint: request
-                    .message
-                    .message_id
-                    .map(|message_id| message_id.to_string()),
-            }
-        }
-
-        fn report_diagnostics(
-            &self,
-            _request: InboxIngressDiagnosticsRequest,
-        ) -> InboxIngressDiagnosticsResponse {
-            InboxIngressDiagnosticsResponse {
-                duplicate_message_ids: 0,
-                messages_without_ids: 0,
-            }
-        }
-    }
-
-    #[derive(Clone)]
-    struct FakeNotificationSink {
-        delivered: Arc<Mutex<Vec<NotificationEvent>>>,
-    }
-
-    impl boundary::sealed::Sealed for FakeNotificationSink {}
-
-    impl NotificationSink for FakeNotificationSink {
-        fn deliver(&self, event: NotificationEvent) -> Result<(), atm_core::error::AtmError> {
-            self.delivered.lock().expect("delivered").push(event);
-            Ok(())
-        }
-    }
-
-    #[test]
-    fn reconcile_runtime_routes_notifications_through_notification_sink_boundary() {
-        let delivered = Arc::new(Mutex::new(Vec::new()));
-        let ingress = FakeInboxIngress::new(vec![InboxIngressImportResponse {
-            source_files: vec![inbox_source_with_message(sample_message(
-                "projected message",
-            ))],
-        }]);
-        let runtime = ReconcileRuntime::new(
-            Arc::new(FakeWatchSource),
-            Arc::new(ingress),
-            Arc::new(FakeNotificationSink {
-                delivered: Arc::clone(&delivered),
-            }),
-        );
-        runtime.start().expect("start");
-
-        let result = runtime.reconcile(request()).expect("reconcile result");
-        assert_eq!(result.observed_paths, 1);
-        assert_eq!(result.imported_sources, 1);
-
-        let delivered = delivered.lock().expect("delivered");
-        assert_eq!(delivered.len(), 1);
-        assert_eq!(
-            delivered[0].kind,
-            atm_core::protocol::NotificationKind::ReconcileComplete
-        );
-
-        runtime.shutdown().expect("shutdown");
-    }
-
-    #[test]
-    fn reconcile_runtime_suppresses_duplicate_notifications_for_same_message_snapshot() {
-        let delivered = Arc::new(Mutex::new(Vec::new()));
-        let watch_polls = Arc::new(AtomicU64::new(0));
-        let repeated_message = sample_message("same logical message");
-        let repeated_source = inbox_source_with_message(repeated_message);
-        let runtime = ReconcileRuntime::new(
-            Arc::new(CountingWatchSource {
-                calls: Arc::clone(&watch_polls),
-            }),
-            Arc::new(FakeInboxIngress::new(vec![
-                InboxIngressImportResponse {
-                    source_files: vec![repeated_source.clone()],
-                },
-                InboxIngressImportResponse {
-                    source_files: vec![repeated_source],
-                },
-            ])),
-            Arc::new(FakeNotificationSink {
-                delivered: Arc::clone(&delivered),
-            }),
-        );
-        runtime.start().expect("start");
-
-        let request = request();
-        let first = runtime.reconcile(request.clone()).expect("first reconcile");
-        let second = runtime.reconcile(request).expect("second reconcile");
-        assert_eq!(first.imported_sources, 1);
-        assert_eq!(second.imported_sources, 1);
-        assert_eq!(watch_polls.load(Ordering::Relaxed), 2);
-
-        let delivered = delivered.lock().expect("delivered");
-        assert_eq!(delivered.len(), 1);
-        assert_eq!(
-            delivered[0].kind,
-            atm_core::protocol::NotificationKind::ReconcileComplete
-        );
-
-        runtime.shutdown().expect("shutdown");
-    }
-
-    #[test]
-    fn reconcile_runtime_bounds_notification_fingerprint_registry_and_re_emits_after_eviction() {
-        let delivered = Arc::new(Mutex::new(Vec::new()));
-        let imports = (0..=MAX_RECONCILE_FINGERPRINT_KEYS)
-            .map(|index| InboxIngressImportResponse {
-                source_files: vec![inbox_source_with_message(sample_message(&format!(
-                    "message-{index}"
-                )))],
-            })
-            .collect::<Vec<_>>();
-        let runtime = ReconcileRuntime::new(
-            Arc::new(FakeWatchSource),
-            Arc::new(FakeInboxIngress::new(imports)),
-            Arc::new(FakeNotificationSink {
-                delivered: Arc::clone(&delivered),
-            }),
-        );
-        runtime.start().expect("start");
-
-        let first_request = request_for("agent-0");
-        runtime
-            .reconcile(first_request.clone())
-            .expect("first reconcile");
-        for index in 1..=MAX_RECONCILE_FINGERPRINT_KEYS {
-            runtime
-                .reconcile(request_for(&format!("agent-{index}")))
-                .expect("bounded reconcile");
-        }
-        runtime
-            .reconcile(first_request)
-            .expect("reconcile after eviction");
-
-        let fingerprints = runtime
-            .inner
-            .notification_fingerprints
-            .lock()
-            .expect("fingerprints");
-        assert_eq!(fingerprints.entries.len(), MAX_RECONCILE_FINGERPRINT_KEYS);
-        assert_eq!(fingerprints.order.len(), MAX_RECONCILE_FINGERPRINT_KEYS);
-        drop(fingerprints);
-
-        let delivered = delivered.lock().expect("delivered");
-        assert_eq!(delivered.len(), MAX_RECONCILE_FINGERPRINT_KEYS + 2);
-
-        runtime.shutdown().expect("shutdown");
-    }
-
-    #[test]
-    fn reconcile_runtime_bounds_per_key_fingerprint_sets() {
-        let runtime = ReconcileRuntime::new(
-            Arc::new(FakeWatchSource),
-            Arc::new(FakeInboxIngress::new(vec![InboxIngressImportResponse {
-                source_files: (0..=MAX_RECONCILE_FINGERPRINTS_PER_KEY)
-                    .map(|index| {
-                        inbox_source_with_message(sample_message(&format!("message-{index}")))
-                    })
-                    .collect(),
-            }])),
-            Arc::new(FakeNotificationSink {
-                delivered: Arc::new(Mutex::new(Vec::new())),
-            }),
-        );
-        runtime.start().expect("start");
-        let request = request();
-        runtime.reconcile(request.clone()).expect("reconcile");
-
-        let fingerprints = runtime
-            .inner
-            .notification_fingerprints
-            .lock()
-            .expect("fingerprints");
-        assert_eq!(
-            fingerprints
-                .entries
-                .get(&ReconcileKey::from_request(&request))
-                .expect("entry")
-                .len(),
-            MAX_RECONCILE_FINGERPRINTS_PER_KEY
-        );
-        drop(fingerprints);
-
-        runtime.shutdown().expect("shutdown");
-    }
-
-    #[test]
-    fn reconcile_runtime_discards_completed_entries_for_timed_out_waiters() {
-        let (started_tx, started_rx) = mpsc::channel();
-        let release: Arc<(Mutex<bool>, Condvar)> = Arc::new((Mutex::new(false), Condvar::new()));
-        let runtime = ReconcileRuntime::new_for_test(
-            Arc::new({
-                let release = Arc::clone(&release);
-                move |_| {
-                    started_tx.send(()).expect("started");
-                    let (released, wake) = &*release;
-                    let mut released = released.lock().expect("released");
-                    while !*released {
-                        let wait = wake
-                            .wait_timeout(released, Duration::from_secs(1))
-                            .expect("wait release");
-                        released = wait.0;
-                    }
-                    Ok(ReconcileResult {
-                        observed_paths: 1,
-                        imported_sources: 1,
-                    })
-                }
-            }),
-            Duration::from_millis(10),
-        );
-        runtime.start().expect("start");
-
-        let runtime_for_thread = runtime.clone();
-        let join = std::thread::spawn(move || runtime_for_thread.reconcile(request()));
-        started_rx
-            .recv_timeout(Duration::from_secs(1))
-            .expect("started");
-        let error = join.join().expect("join").expect_err("timeout should fail");
-        assert!(error.message.contains("timed out"));
-
-        let (released, wake) = &*release;
-        *released.lock().expect("released") = true;
-        wake.notify_all();
-        assert!(
-            runtime.wait_for_pending_count_for_test(0, Duration::from_secs(1)),
-            "pending reconcile work did not drain after release"
-        );
-        assert_eq!(runtime.state_counts_for_test(), (0, 0, 0));
-
-        runtime.shutdown().expect("shutdown");
-    }
-
-    #[derive(Clone)]
-    struct CountingWatchSource {
-        calls: Arc<AtomicU64>,
-    }
-
-    impl boundary::sealed::Sealed for CountingWatchSource {}
-
-    impl WatchEventSource for CountingWatchSource {
-        fn poll(
-            &self,
-            _request: WatchSubscriptionRequest,
-        ) -> Result<WatchEventBatch, atm_core::error::AtmError> {
-            self.calls.fetch_add(1, Ordering::Relaxed);
-            Ok(WatchEventBatch {
-                paths: vec![std::env::temp_dir().join("watch.json")],
-            })
-        }
-    }
-
-    fn inbox_source_with_message(
-        message: MessageEnvelope,
-    ) -> atm_core::boundary::InboxSourceFileRecord {
-        atm_core::boundary::InboxSourceFileRecord {
-            path: std::env::temp_dir().join("watch.json"),
-            messages: vec![message],
-        }
-    }
-
-    fn sample_message(text: &str) -> MessageEnvelope {
-        let message_id = AtmMessageId::new();
-
-        MessageEnvelope {
-            from: ROLE_TEAM_LEAD.parse().expect("agent"),
-            text: text.to_string(),
-            timestamp: IsoTimestamp::from_datetime(Utc::now()),
-            read: false,
-            source_team: Some("test-team".parse().expect("team")),
-            summary: Some("summary".to_string()),
-            message_id: Some(message_id),
-            pending_ack_at: None,
-            acknowledged_at: None,
-            acknowledges_message_id: None,
-            parent_message_id: None,
-            thread_mode: None,
-            expires_at: None,
-            task_id: None,
-            extra: Map::new(),
-        }
+#[cfg(test)]
+impl NotificationSink for TestNotificationSink {
+    fn deliver(&self, _event: NotificationEvent) -> Result<(), AtmError> {
+        Ok(())
     }
 }
+
+#[cfg(test)]
+#[path = "reconcile_runtime_tests.rs"]
+mod tests;
