+++ docs/phase-Ye/sprint-Y22.md
@@ -1,7 +1,7 @@
 ---
 id: Y.22
 title: Reconcile Runtime Cutover
-status: planned
+status: complete
 branch: feature/pYe-s22-reconcile-runtime-cutover
 worktree: ../atm-core-worktrees/feature/pYe-s22-reconcile-runtime-cutover
 target: integrate/phase-Y
@@ -54,7 +54,6 @@ ADR-015 ownership in this sprint:
 - `crates/atm-daemon/Cargo.toml`
 - `crates/atm-daemon/src/worker_support.rs`
 - `crates/atm-daemon/src/reconcile_runtime.rs`
-- `crates/atm-daemon/src/composition.rs`
 - `docs/adr/ADR-015-daemon-runtime-snapshot-and-worker-ownership.md`
 - `docs/atm-daemon/requirements.md`
 - `docs/atm-daemon/architecture.md`
@@ -110,7 +109,7 @@ and does not reintroduce a lane-local duplicate.
 2. worker coalesces and debounces
 3. worker executes watch/ingress/notification work
 4. worker replies to all waiters and updates reconcile runtime status
-5. shutdown sends one control command and proves bounded termination
+5. shutdown flips explicit runtime status and proves bounded worker termination
 
 ## Deliverables
 
@@ -120,6 +119,8 @@ and does not reintroduce a lane-local duplicate.
   actor state; there is no surviving side mutex for fingerprint ownership
 - reconcile coalescing, completion fanout, and bounded shutdown are proven on
   the final actor design
+- `JoinHandleOwner` remains the only narrow mutex helper on the reconcile lane,
+  and it owns join lifecycle only
 - daemon requirements, architecture, and boundary docs reflect the cutover
   implementation shape for reconcile ownership
